+++ projects/packages/my-jetpack/_inc/components/product-detail-card/index.jsx
@@ -6,17 +6,7 @@ import classnames from 'classnames';
 import { Icon, check, plus } from '@wordpress/icons';
 import { getCurrencyObject } from '@automattic/format-currency';
 import { __, sprintf } from '@wordpress/i18n';
-import {
-	AntiSpamIcon,
-	BackupIcon,
-	CheckmarkIcon,
-	Col,
-	getIconBySlug,
-	Container,
-	ScanIcon,
-	StarIcon,
-	Text,
-} from '@automattic/jetpack-components';
+import { CheckmarkIcon, getIconBySlug, StarIcon, Text, H3 } from '@automattic/jetpack-components';
 
 /**
  * Internal dependencies
@@ -24,32 +14,8 @@ import {
 import styles from './style.module.scss';
 import getProductCheckoutUrl from '../../utils/get-product-checkout-url';
 import useMyJetpackConnection from '../../hooks/use-my-jetpack-connection';
-import { useProduct } from '../../hooks/use-product';
-
 import ProductDetailButton from '../product-detail-button';
-
-/**
- * Simple react component to render the product icon,
- * depending on the product slug.
- *
- * @param {string} slug - The product slug.
- * @returns {object}    ProductDetailCard react component.
- */
-function ProductIcon( { slug } ) {
-	switch ( slug ) {
-		case 'anti-spam':
-			return <AntiSpamIcon size={ 24 } />;
-
-		case 'backup':
-			return <BackupIcon size={ 24 } />;
-
-		case 'scan':
-			return <ScanIcon size={ 24 } />;
-
-		default:
-			return null;
-	}
-}
+import { useProduct } from '../../hooks/use-product';
 
 /**
  * React component to render the price.
@@ -156,91 +122,96 @@ const ProductDetailCard = ( { slug, onClick, trackButtonClick, className } ) => 
 		}
 	}, [ onClick, trackButtonClick ] );
 
+	/**
+	 * Temporary ProductIcon component.
+	 * Todo: Fix in product-icons component.
+	 * https://github.com/Automattic/jetpack/issues/23640
+	 *
+	 * @param {object} props      - Component props.
+	 * @param {string} props.slug - Product icon slug
+	 * @returns {object}            Icon Product component.
+	 */
+	function ProductIcon( { slug: iconSlug } ) {
+		const ProIcon = getIconBySlug( iconSlug );
+		if ( ! ProIcon ) {
+			return () => null;
+		}
+
+		return (
+			<div className={ styles[ 'product-icon' ] }>
+				<ProIcon />
+			</div>
+		);
+	}
+
 	return (
-		<Container
-			className={ classnames( styles.card, className ) }
-			fluid
-			horizontalGap={ 0 }
-			horizontalSpacing={ 0 }
+		<div
+			className={ classnames( styles.card, className, {
+				[ styles[ 'is-bundle-card' ] ]: isBundle,
+			} ) }
 		>
 			{ isBundle && (
-				<Col className={ styles[ 'card-header' ] }>
+				<div className={ styles[ 'card-header' ] }>
 					<StarIcon className={ styles[ 'product-bundle-icon' ] } size={ 16 } />
 					<Text variant="label">{ __( 'Popular upgrade', 'jetpack-my-jetpack' ) }</Text>
-				</Col>
+				</div>
 			) }
 
-			<Col>
-				<Container horizontalSpacing={ 5 } horizontalGap={ 2 } className={ styles.container }>
-					<Col>
-						{ isBundle && <div>{ icons }</div> }
+			<div className={ styles.container }>
+				{ isBundle && <div className={ styles[ 'product-bundle-icons' ] }>{ icons }</div> }
+				<ProductIcon slug={ slug } />
 
-						<ProductIcon slug={ slug } />
-					</Col>
+				<H3>{ title }</H3>
+				<Text mb={ 3 }>{ longDescription }</Text>
 
-					<Text variant="headline-small" className={ styles.title } component={ Col }>
-						{ title }
-					</Text>
+				<ul className={ styles.features }>
+					{ features.map( ( feature, id ) => (
+						<Text component="li" key={ `feature-${ id }` } variant="body">
+							<Icon icon={ check } size={ 24 } />
+							{ feature }
+						</Text>
+					) ) }
+				</ul>
 
-					<Text className={ styles.name } component={ Col }>
-						{ longDescription }
-					</Text>
-
-					<Col>
-						<ul className={ styles.features }>
-							{ features.map( ( feature, id ) => (
-								<Text component="li" key={ `feature-${ id }` } variant="body-extra-small">
-									<Icon icon={ check } size={ 30 } />
-									{ feature }
-								</Text>
-							) ) }
-						</ul>
-					</Col>
-
-					{ needsPurchase && (
-						<Col className={ styles[ 'price-container' ] }>
+				{ needsPurchase && (
+					<>
+						<div className={ styles[ 'price-container' ] }>
 							<Price value={ price } currency={ currencyCode } isOld={ true } />
 							<Price value={ discountPrice } currency={ currencyCode } isOld={ false } />
-							<Text className={ styles[ 'price-description' ] }>
-								{ __( '/month, paid yearly', 'jetpack-my-jetpack' ) }
-							</Text>
-						</Col>
-					) }
-
-					{ isFree && (
-						<Text variant="title-small" component={ Col }>
-							{ __( 'Free', 'jetpack-my-jetpack' ) }
+						</div>
+						<Text className={ styles[ 'price-description' ] }>
+							{ __( '/month, paid yearly', 'jetpack-my-jetpack' ) }
 						</Text>
-					) }
+					</>
+				) }
 
-					<Col>
-						{ ( ! isBundle || ( isBundle && ! hasRequiredPlan ) ) && (
-							<Text
-								component={ ProductDetailButton }
-								onClick={ clickHandler }
-								isLoading={ isFetching }
-								isPrimary={ ! isBundle }
-								href={ onClick ? undefined : addProductUrl }
-								className={ styles[ 'checkout-button' ] }
-								variant="body"
-							>
-								{
-									/* translators: placeholder is product name. */
-									sprintf( __( 'Add %s', 'jetpack-my-jetpack' ), title )
-								}
-							</Text>
-						) }
+				{ isFree && <H3>{ __( 'Free', 'jetpack-my-jetpack' ) }</H3> }
 
-						{ isBundle && hasRequiredPlan && (
-							<div className={ styles[ 'product-has-required-plan' ] }>
-								<CheckmarkIcon size={ 36 } />
-								<Text>{ __( 'Active on your site', 'jetpack-my-jetpack' ) }</Text>
-							</div>
-						) }
-					</Col>
-				</Container>
-			</Col>
-		</Container>
+				{ ( ! isBundle || ( isBundle && ! hasRequiredPlan ) ) && (
+					<Text
+						component={ ProductDetailButton }
+						onClick={ clickHandler }
+						isLoading={ isFetching }
+						isPrimary={ ! isBundle }
+						href={ onClick ? undefined : addProductUrl }
+						className={ styles[ 'checkout-button' ] }
+						variant="body"
+					>
+						{
+							/* translators: placeholder is product name. */
+							sprintf( __( 'Add %s', 'jetpack-my-jetpack' ), title )
+						}
+					</Text>
+				) }
+
+				{ isBundle && hasRequiredPlan && (
+					<div className={ styles[ 'product-has-required-plan' ] }>
+						<CheckmarkIcon size={ 36 } />
+						<Text>{ __( 'Active on your site', 'jetpack-my-jetpack' ) }</Text>
+					</div>
+				) }
+			</div>
+		</div>
 	);
 };