+++ projects/plugins/jetpack/extensions/shared/components/product-management-controls/use-open-block-sidebar.js
@@ -1,22 +1,27 @@
 /**
  * WordPress dependencies
  */
-import { useDispatch, select } from '@wordpress/data';
+import { store as blockEditorStore } from '@wordpress/block-editor';
+import { useDispatch } from '@wordpress/data';
+import { getEditorType, SITE_EDITOR, WIDGET_EDITOR } from '../../get-editor-type';
 
-export default function useOpenBlockSidebar() {
+export default function useOpenBlockSidebar( clientId ) {
+	const editorType = getEditorType();
+	const { selectBlock } = useDispatch( blockEditorStore );
 	const { enableComplementaryArea } = useDispatch( 'core/interface' );
-	const isSiteEditor = !! select( 'core/edit-site' );
-	const isWidgetsEditor = !! select( 'core/edit-widgets' );
 
 	return () => {
-		if ( isSiteEditor ) {
-			enableComplementaryArea( 'core/edit-site', 'edit-site/block-inspector' );
-			return;
+		if ( clientId ) {
+			selectBlock( clientId );
 		}
-		if ( isWidgetsEditor ) {
-			enableComplementaryArea( 'core/edit-widgets', 'edit-widgets/block-inspector' );
-			return;
+
+		switch ( editorType ) {
+			case SITE_EDITOR:
+				return enableComplementaryArea( 'core/edit-site', 'edit-site/block-inspector' );
+			case WIDGET_EDITOR:
+				return enableComplementaryArea( 'core/edit-widgets', 'edit-widgets/block-inspector' );
+			default:
+				enableComplementaryArea( 'core/edit-post', 'edit-post/block' );
 		}
-		enableComplementaryArea( 'core/edit-post', 'edit-post/block' );
 	};
 }