+++ projects/js-packages/connection/CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ### This is a list detailing changes for the Jetpack RNA Connection Component releases.
 
+## 0.17.7 - 2022-04-05
+### Added
+- Add skipUserConnection option to connectScreen component
+
+### Changed
+- Updated package dependencies.
+
+## 0.17.6 - 2022-03-31
+### Changed
+- My Jetpack: tweak connection card styles and wording
+
 ## 0.17.5 - 2022-03-29
 ### Added
 - Connection: Use heading components to render status card cmp